JO-4: added faculty update page; restructured faculty app
