+++ db/models.py
@@ -1,26 +1,5 @@
-from typing import Annotated, List, Optional
 
-from fastapi import Depends, FastAPI, HTTPException, Query
-
-from sqlmodel import Field, Session, SQLModel, Relationship, create_engine, select
-
-
-DB_HOST = 'localhost'
-DB_PORT = 5432
-DB_NAME = 'students'
-DB_USER = 'postgres'
-DB_PASSWORD = 'root'
-
-engine = create_engine(f"postgresql+psycopg2://{DB_USER}:{DB_PASSWORD}@{DB_HOST}:{DB_PORT}/{DB_NAME}")
-
-
-def get_session():
-    with Session(engine) as session:
-        yield session
-
-
-SessionDep = Annotated[Session, Depends(get_session)]
-
+from sqlmodel import Field, SQLModel, Relationship
 
 
 class Faculty(SQLModel, table=True):
@@ -91,4 +70,3 @@ class Teacher(SQLModel, table=True):
 
     def __repr__(self) -> str:
         return f"Teacher(id={self.id!r}, first_name={self.first_name!r}, middle_name={self.middle_name!r}, last_name={self.last_name!r}"
-
