+++ faculties/__init__.py
@@ -1,0 +1,2 @@
+from .api import faculties_api_router
+from .base import faculties_router
