+++ local_settings.py.template
@@ -1,0 +1,5 @@
+DB_HOST = 'localhost'
+DB_PORT = 5432
+DB_NAME = 'students'
+DB_USER = 'postgres'
+DB_PASSWORD = 'root'
