+++ templates/_base/500.html
