+++ templates/_base/layout-static.html
