+++ templates/faculties/faculties.html
@@ -13,7 +13,7 @@
                     { data: 'id',
                       width: 100,
                       render: function ( data, type, row) {
-                        return '<a href="/faculty/' + data + '">' + data + '</a>';
+                        return '<a href="/faculties/' + data + '/edit/">' + data + '</a>';
                       }
                     },
                     { data: 'name', className: 'dt-body-right' },
