+++ templates/faculties/faculty_edit.html
@@ -1,0 +1,45 @@
+{% extends "base.html" %}
+
+{% block header %}
+<h1 class="mt-4">Факультет {{ faculty.name }}</h1>
+{% endblock %}
+
+{% block breadcrumbs %}
+<ol class="breadcrumb mb-4">
+    <li class="breadcrumb-item active">Факультеты</li>
+    <li class="breadcrumb-item active">{{ faculty.name }}</li>
+</ol>
+{% endblock %}
+
+{% block content %}
+
+<div class="container">
+    <div class="row justify-content-center">
+        <div class="col-lg-5">
+            <div class="card shadow-lg border-0 rounded-lg mt-5">
+                <div class="card-body">
+                    <form action="{{ url_for('faculty_update', faculty_id=faculty.id) }}" method="post">
+                        <div class="row mb-3">
+                            <div class="col-md-12">
+                                <div class="form-floating mb-3">
+                                    <input class="form-control" id="name" name="name" type="text" placeholder="Название" value="{{ faculty.name }}" />
+                                    <label for="name">Название</label>
+                                </div>
+                                <div class="form-floating">
+                                    <input class="form-control" id="num" name="num" type="number" placeholder="Номер" value="{{ faculty.num }}" />
+                                    <label for="num">Номер</label>
+                                </div>
+                            </div>
+                        </div>
+                        <div class="row">
+                            <div class="col-md-4"><button class="btn btn-primary" type="submit">Сохранить</button></div>
+                            <div class="col-md-4 px-0"><a class="btn btn-secondary" href="{{ url_for('faculty_details', faculty_id=faculty.id) }}">Отменить</a></div>
+<!--                            <div class="col-md-4"><button class="btn btn-secondary" type="button" onclick="history.back()">Отменить</button></div>-->
+                        </div>
+                    </form>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+{% endblock %}
